Extract utilities to own file
